+++ src/app/components/header/header.ts
@@ -1,8 +1,9 @@
 import { Component } from '@angular/core';
+import { Button } from '../ui/button/button';
 
 @Component({
   selector: 'header[app-header]',
-  imports: [],
+  imports: [Button],
   templateUrl: './header.html',
   styleUrl: './header.scss'
 })