+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.4.0] - 2025-06-21
+### Main
+ - Added a notification service
+ - Display a notification
+ - Display de notification with 2 choice buttons
+
+
 ## [1.3.0] - 2025-06-19
 ### Main
 - Added an authentication service